Successfully read controller string with HID i/o
== DETAILS
I've confirmed this works with a switch pro controller.

It does *not* presently work with a dualsense controller, but
that's because the ds3 has a proprietary activation packet that
needs to be sent that I haven't implemented here.

The bytes returned are technically unicode and I'm doing a
poor man's ASCII conversion here.